getredash/redash#2117 Revisit Plotly integration
