+++ client/app/visualizations/chart/chart-editor.html
@@ -174,11 +174,6 @@
         <i class="input-helper"></i> Show Labels
       </label>
     </div>
-
-    <div class="form-group">
-      <label class="control-label">Height</label>
-      <input name="x-axis-height" type="number" class="form-control" ng-model="options.bottomMargin">
-    </div>
   </div>
 
   <div ng-show="currentTab == 'yAxis'">
